// src/API/AutoShopApi/models/CreateUpdateUserDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AutoShopAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateUpdateUserDTO
 */
export interface CreateUpdateUserDTO {
    /**
     * 
     * @type {string}
     * @memberof CreateUpdateUserDTO
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateUpdateUserDTO
     */
    email?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateUpdateUserDTO
     */
    password?: string | null;
    /**
     * 
     * @type {number}
     * @memberof CreateUpdateUserDTO
     */
    carId?: number | null;
}

/**
 * Check if a given object implements the CreateUpdateUserDTO interface.
 */
export function instanceOfCreateUpdateUserDTO(value: object): value is CreateUpdateUserDTO {
    return true;
}

export function CreateUpdateUserDTOFromJSON(json: any): CreateUpdateUserDTO {
    return CreateUpdateUserDTOFromJSONTyped(json, false);
}

export function CreateUpdateUserDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateUpdateUserDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'email': json['email'] == null ? undefined : json['email'],
        'password': json['password'] == null ? undefined : json['password'],
        'carId': json['carId'] == null ? undefined : json['carId'],
    };
}

export function CreateUpdateUserDTOToJSON(json: any): CreateUpdateUserDTO {
    return CreateUpdateUserDTOToJSONTyped(json, false);
}

export function CreateUpdateUserDTOToJSONTyped(value?: CreateUpdateUserDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'email': value['email'],
        'password': value['password'],
        'carId': value['carId'],
    };
}